backend_for_project/requirements.txt
fastapi
uvicorn[standard]
opencv-python
scikit-learn
numpy
supabase
python-dotenv
python-multipart
httpx
